Commands para para las tareas programadas

+++ app/Http/Controllers/ScheduleController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 use App\Schedule;
 use App\Program;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Log;
 
 class ScheduleController extends Controller
 {
@@ -90,36 +91,45 @@ class ScheduleController extends Controller
             ->where('schedule_id',$id)
             ->delete();
 
-            // Agregar programas del horario
-            for ($i = 1; $i <= 7; $i++) {
-                switch (count($request->todos[$i - 1])) {
-                    case 0:
-                        break;
-                    case 2:
-                        $program = new Program();
-                        $program->entrada1 = $request->todos[$i - 1][0];
-                        $program->salida1 = $request->todos[$i - 1][1];
-                        $program->dia_id = $i;
-                        $program->schedule_id = $horario->id;
-                        $program->save();
-                        break;
-                    case 4:
-                        $program = new Program();
-                        $program->entrada1 = $request->todos[$i - 1][0];
-                        $program->salida1 = $request->todos[$i - 1][1];
-                        $program->entrada2 = $request->todos[$i - 1][2];
-                        $program->salida2 = $request->todos[$i - 1][3];
-                        $program->dia_id = $i;
-                        $program->schedule_id = $horario->id;
-                        $program->save();
-                        break;
-                    default:
-                        break;
+            $cant_reg_borr = DB::table('programs')
+            ->where('schedule_id',$id)
+            ->get();
+
+            if(count($cant_reg_borr) == 0){
+
+                // Agregar programas del horario
+                for ($i = 1; $i <= 7; $i++) {
+                    switch (count($request->todos[$i - 1])) {
+                        case 0:
+                            break;
+                        case 2:
+                            $program = new Program();
+                            $program->entrada1 = $request->todos[$i - 1][0];
+                            $program->salida1 = $request->todos[$i - 1][1];
+                            $program->dia_id = $i;
+                            $program->schedule_id = $horario->id;
+                            $program->save();
+                            break;
+                        case 4:
+                            $program = new Program();
+                            $program->entrada1 = $request->todos[$i - 1][0];
+                            $program->salida1 = $request->todos[$i - 1][1];
+                            $program->entrada2 = $request->todos[$i - 1][2];
+                            $program->salida2 = $request->todos[$i - 1][3];
+                            $program->dia_id = $i;
+                            $program->schedule_id = $horario->id;
+                            $program->save();
+                            break;
+                        default:
+                            break;
+                    }
                 }
+
+                $horario->save = true;
+                return $horario;
+
             }
 
-            $horario->save = true;
-            return $horario;
         }
         
         $horario->save = false;
@@ -128,7 +138,15 @@ class ScheduleController extends Controller
 
     public function destroy($id)
     {
-        $schedule = Schedule::find($id);
-        $schedule->delete();
+        $cant_reg = DB::table('relationships')
+        ->where('schedule_id',$id)
+        ->get();
+        
+        if(count($cant_reg) == 0){
+            $schedule = Schedule::find($id);
+            $schedule->delete();
+        }
+
+        Log::info("Existen relaciones con el horario que desea eliminar. Borre estás y podrá eliminar el horario.");
     }
 }
